requirements.txt
argparse==1.2.1
lxml==3.4.4
PyYAML==3.11
requests==2.4.3
